Refactor of folderPicker.
